[#784] Move shared field components to own file
